Process copies rather than backed-up originals
Appends a suffix (currently "formatted") to copies of all of the source files, does the Python processing on these copies, and then runs xetex, etc.

main.tex was updated so it now includes the .tex.formatted versions.

To accommodate this, method replaceVariants() of utils.py was modified so that it takes a collection of file-filename pairs, and main() was changed so that if there are filenames given as command line arguments, replaceVariants will be run on file-filename pairs from these filenames.